refactor: rename ollama package to ai for clarity in client usage
